Improving setup files for easier build
The new setup.cfg file contains the installation info.  That makes
setup.py much simpler.  The mastcasjob.py is moved into a subdirectory
and renamed __init__.py to work with this build configuration.

The big advantage of this is that it should automatically get a current
version of the casjobs.py module from github rather than relying on
complicated instructions in the readme file.

Update documentation

The installation instructions are a little simpler now that the casjobs
module dependence is handled better.  The example has been updated to
Python 3 (and a comment added about Python 2.7 compatibility).  Some
additional documentation has been added about other functions beyond the
quick example.

Update version number to 0.0.4

A new version number seems certainly justified by the combination of the new
installation approach and the requirement the installation imposes of Python
3.5