Fix bug in fast_table when table names include '.'
Table names including '.' did not work correctly in fast_table.  Such
names are rare (and can cause problems in the Casjobs web interface as
well -- e.g., they do not appear in the MyDB list even though they are
present).  But with the fix (and an associated fix in the
quick_casjobs.cgi service), such tables now work.