Work around changed astropy behavior
astropy.fits now interprets the FITS tables generated by Casjobs as
consisting of all columns being arrays with shape (1,).  This happens
because there is a TDIMn = 1 entry in the header.  They say it is a
problem with the FITS files and refuse to fix it.

Since we can't change the FITS files generated by Casjobs, we fix it in
code in the get_table() function.

mastcasjobs/__init__.py: Call .squeeze() to remove final dimension from
columns if it is unity.

setup.cfg: Bump the version number up to 0.0.5