Fix for bad FITS table issues
The FITS tables generated by MAST Casjobs have some messy problems that
are the result of incorrectly specified TDIM keywords.  Older versions
of astropy.io.fits worked around these problems and produced good
results.  But the current version produces terrible results, with all
columns appearing as arrays of dimension (1,), and with string columns
truncated to just a single character.

While that behavior in astropy.io.fits ought to be improved, the core
problem is that the TDIM keywords in these files are wrong.  We know
that database queries will not produce any array-valued columns.  So the
fix is to delete all the TDIM columns in the FITS header before
converting the data to a table.