Make POST default instead of GET
Make request_type='POST' the default in the MastCasJobs.__init__ method.
There is no real downside to this, and it avoids problems that appear
for long query strings when using GET.

Add a custom version of _parse_error() that pulls out some status info
when URLs are blocked for some reason.  This should help debugging
firewall issues in the future.

Add a release notes section in the readme with short summaries of the
changes betwen versions.

Bump the version up to 0.0.7.